% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{A start for a function}
\usage{
myf(x)
}
\arguments{
\item{x}{vector of data}
}
\value{
a vector of the elements of x squared
}
\description{
This is a start for math4773
}
\examples{
myf(1:10)
}
